requirements.txt
weasyprint==52.2
pybars3==0.9.7
pyyaml==5.3.1
click==7.1.2